Added note about newlines in PING/PONG section
I have set up multiple adapters based on the information in this readme, and multiple times I have gone down a long road of trying to understand why the Agent isn't receiving my pongs, only to discover its because I didn't provide a newline at the end. So I think it might be good to add this little note to that section so its more clear.